[REFACTOR] ensure compliance with SOLID principles
